Moved hexo specific stuff into module
