+++ anemetrohexo/__init__.py
@@ -1,0 +1,5 @@
+"""
+  Outer Imports
+"""
+
+from anemetrohexo.hexo import *
